fix: move affiliate utils into src/ to fix API build
The API tsconfig has rootDir:"src", so importing from tools/static-site/utils.ts
caused TS6059. Move affiliate types and functions to src/shared/affiliate-utils.ts
and re-export from tools/static-site/utils.ts for backward compatibility.
